Set up controllers and modify views to handle updating authors
Modify author_create_post to start the creation of a new author instance and return that to the form view if there are errors. This is so that the view can take advantage of the schema's virtuals when rendering the template.
Modify author_detail.pug to include link to update an author.